Mrs Baylock thinks unholy loch refreshes are hellish
